Fix a bad merge. (#72)
